+++ forui/lib/foundation.dart
@@ -2,6 +2,7 @@
 library forui.foundation;
 
 export 'src/foundation/border_radius.dart';
+export 'src/foundation/collapsible.dart';
 export 'src/foundation/doc_templates.dart' hide Focus, Scroll, Semantics, WidgetStates;
 export 'src/foundation/focused_outline.dart';
 export 'src/foundation/form_field_properties.dart';
