+++ forui/lib/src/widgets/accordion/accordion_item.dart
@@ -5,6 +5,7 @@ import 'package:flutter/rendering.dart';
 import 'package:flutter/widgets.dart';
 
 import 'package:forui/forui.dart';
+import 'package:forui/src/foundation/collapsible.dart';
 import 'package:forui/src/widgets/accordion/accordion.dart';
 import 'package:forui/src/widgets/accordion/accordion_controller.dart';
 
@@ -140,17 +141,11 @@ class _FAccordionItemState extends State<FAccordionItem> with TickerProviderStat
                       ),
                     ),
               ),
-              // We use a combination of a custom render box & clip rect to avoid visual oddities. This is caused by
-              // RenderPaddings (created by Paddings in the child) shrinking the constraints by the given padding, causing the
-              // child to layout at a smaller size while the amount of padding remains the same.
-              _Expandable(
+              FCollapsible(
                 value: _controller!.value,
-                child: ClipRect(
-                  clipper: _Clipper(_controller!.value),
-                  child: Padding(
-                    padding: style.childPadding,
-                    child: DefaultTextStyle(style: style.childTextStyle, child: widget.child),
-                  ),
+                child: Padding(
+                  padding: style.childPadding,
+                  child: DefaultTextStyle(style: style.childTextStyle, child: widget.child),
                 ),
               ),
               FDivider(style: style.dividerStyle),
@@ -165,84 +160,4 @@ class _FAccordionItemState extends State<FAccordionItem> with TickerProviderStat
     _controller?.dispose();
     super.dispose();
   }
-}
-
-class _Expandable extends SingleChildRenderObjectWidget {
-  final double value;
-
-  const _Expandable({required this.value, required super.child});
-
-  @override
-  RenderObject createRenderObject(BuildContext _) => _RenderExpandable(value);
-
-  @override
-  void updateRenderObject(BuildContext context, _RenderExpandable renderObject) => renderObject..value = value;
-
-  @override
-  void debugFillProperties(DiagnosticPropertiesBuilder properties) {
-    super.debugFillProperties(properties);
-    properties.add(PercentProperty('value', value));
-  }
-}
-
-class _RenderExpandable extends RenderBox with RenderObjectWithChildMixin<RenderBox> {
-  double _value;
-
-  _RenderExpandable(this._value);
-
-  @override
-  void performLayout() {
-    if (child case final child?) {
-      child.layout(constraints.normalize(), parentUsesSize: true);
-      size = Size(child.size.width, child.size.height * _value);
-    } else {
-      size = constraints.smallest;
-    }
-  }
-
-  @override
-  void paint(PaintingContext context, Offset offset) {
-    if (child case final child?) {
-      context.paintChild(child, offset);
-    }
-  }
-
-  @override
-  bool hitTest(BoxHitTestResult result, {required Offset position}) {
-    if (size.contains(position) && child!.hitTest(result, position: position)) {
-      result.add(BoxHitTestEntry(this, position));
-      return true;
-    }
-
-    return false;
-  }
-
-  double get value => _value;
-
-  set value(double value) {
-    if (_value == value) {
-      return;
-    }
-
-    _value = value;
-    markNeedsLayout();
-  }
-
-  @override
-  void debugFillProperties(DiagnosticPropertiesBuilder properties) {
-    super.debugFillProperties(properties);
-    properties.add(PercentProperty('value', value));
-  }
-}
-
-class _Clipper extends CustomClipper<Rect> {
-  final double percentage;
-
-  _Clipper(this.percentage);
-
-  @override
-  Rect getClip(Size size) => Offset.zero & Size(size.width, size.height * percentage);
-
-  @override
-  bool shouldReclip(covariant _Clipper oldClipper) => oldClipper.percentage != percentage;
 }
